rename solution, projects, name spaces
