+++ Nicks.UnitTester.Shared/CodeParser.cs
@@ -3,7 +3,7 @@ using System.Linq;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 
-namespace Nicks.UnitTester.Shared
+namespace CloudExternals.TestGen.Shared
 {
     public class CodeParser
     {
