+++ Nicks.UnitTester.Shared/FileParseResult.cs
@@ -1,4 +1,4 @@
-﻿namespace Nicks.UnitTester.Shared
+﻿namespace CloudExternals.TestGen.Shared
 {
     public class FileParseResult : CodeParseResult
     {
