+++ Nicks.UnitTester.Tests/CloudExternals.TestGen.Tests.csproj
@@ -22,7 +22,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\Nicks.UnitTester.Shared\Nicks.UnitTester.Shared.csproj" />
+    <ProjectReference Include="..\Nicks.UnitTester.Shared\CloudExternals.TestGen.Shared.csproj" />
   </ItemGroup>
 
 </Project>
